requirements.txt
autopep8==2.0.1
cachetools==5.3.0
colorama==0.4.6
pefile==2023.2.7
pycodestyle==2.10.0
ruamel.yaml
tqdm==4.64.1
